Crate adexchangebuyer1d3

Features

Handle the following Resources with ease from the central hub ...

accounts (get, list, patch and update)

billing info (get and list)

budget (get, patch and update)

creatives (get, insert and list)

direct deals (get and list)

performance report (list)

pretargeting config (delete, get, insert, list, patch and update)

Everything else about the Ad Exchange Buyer API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.pretargeting_config().insert(...).doit()
let r = hub.pretargeting_config().patch(...).doit()
let r = hub.pretargeting_config().get(...).doit()
let r = hub.pretargeting_config().update(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate adexchangebuyer1d3;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeBuyer::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: PretargetingConfig = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.pretargeting_config().patch(&req, "accountId", "configId")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Account
Configuration data for an Ad Exchange buyer account.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update (request|response)

get (response)

list (none)

patch (request|response)

AccountBidderLocation
Your bidder locations that have distinct URLs.

This type is not used in any activity, and only used as part of another schema.

AccountGetMethodBuilder
Gets one account by ID.

AccountListMethodBuilder
Retrieves the authenticated user's list of accounts.

AccountMethodsBuilder
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdExchangeBuyer hub.

AccountPatchMethodBuilder
Updates an existing account. This method supports patch semantics.

AccountUpdateMethodBuilder
Updates an existing account.

AccountsList
An account feed lists Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single buyer account.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

AdExchangeBuyer
Central instance to access all AdExchangeBuyer related resource activities

BillingInfo
The configuration data for an Ad Exchange billing info.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

BillingInfoGetMethodBuilder
Returns the billing information for one account specified by account ID.

BillingInfoList
A billing info feed lists Billing Info the Ad Exchange buyer account has access to. Each entry in the feed corresponds to a single billing info.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

BillingInfoListMethodBuilder
Retrieves a list of billing information for all accounts of the authenticated user.

BillingInfoMethodsBuilder
A builder providing access to all methods supported on billingInfo resources. It is not used directly, but through the AdExchangeBuyer hub.

Budget
The configuration data for Ad Exchange RTB - Budget API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch (request|response)

update (request|response)

get (response)

BudgetGetMethodBuilder
Returns the budget information for the adgroup specified by the accountId and billingId.

BudgetMethodsBuilder
A builder providing access to all methods supported on budget resources. It is not used directly, but through the AdExchangeBuyer hub.

BudgetPatchMethodBuilder
Updates the budget amount for the budget of the adgroup specified by the accountId and billingId, with the budget amount in the request. This method supports patch semantics.

BudgetUpdateMethodBuilder
Updates the budget amount for the budget of the adgroup specified by the accountId and billingId, with the budget amount in the request.

Creative
A creative and its classification data.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert (request|response)

list (none)

get (response)

CreativeCorrections
Shows any corrections that were applied to this creative. Read-only. This field should not be set in requests.

This type is not used in any activity, and only used as part of another schema.

CreativeDisapprovalReasons
The reasons for disapproval, if any. Note that not all disapproval reasons may be categorized, so it is possible for the creative to have a status of DISAPPROVED with an empty list for disapproval_reasons. In this case, please reach out to your TAM to help debug the issue. Read-only. This field should not be set in requests.

This type is not used in any activity, and only used as part of another schema.

CreativeFilteringReasons
The filtering reasons for the creative. If this feature is not enabled, please ask your technical account manager. Read-only. This field should not be set in requests.

This type is not used in any activity, and only used as part of another schema.

CreativeFilteringReasonsReasons
The filtering reasons.

This type is not used in any activity, and only used as part of another schema.

CreativeGetMethodBuilder
Gets the status for a single creative. A creative will be available 30-40 minutes after submission.

CreativeInsertMethodBuilder
Submit a new creative.

CreativeListMethodBuilder
Retrieves a list of the authenticated user's active creatives. A creative will be available 30-40 minutes after submission.

CreativeMethodsBuilder
A builder providing access to all methods supported on creative resources. It is not used directly, but through the AdExchangeBuyer hub.

CreativesList
The creatives feed lists the active creatives for the Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single creative.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

DirectDeal
The configuration data for an Ad Exchange direct deal.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

list (none)

DirectDealGetMethodBuilder
Gets one direct deal by ID.

DirectDealListMethodBuilder
Retrieves the authenticated user's list of direct deals.

DirectDealMethodsBuilder
A builder providing access to all methods supported on directDeal resources. It is not used directly, but through the AdExchangeBuyer hub.

DirectDealsList
A direct deals feed lists Direct Deals the Ad Exchange buyer account has access to. This includes direct deals set up for the buyer account as well as its merged stream seats.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

PerformanceReport
The configuration data for an Ad Exchange performance report list.

This type is not used in any activity, and only used as part of another schema.

PerformanceReportList
The configuration data for an Ad Exchange performance report list. https://sites.google.com/a/google.com/adx-integration/Home/engineering/binary-releases/rtb-api-release https://cs.corp.google.com/#piper///depot/google3/contentads/adx/tools/rtb_api/adxrtb.py

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

PerformanceReportListMethodBuilder
Retrieves the authenticated user's list of performance metrics.

PerformanceReportMethodsBuilder
A builder providing access to all methods supported on performanceReport resources. It is not used directly, but through the AdExchangeBuyer hub.

PretargetingConfig
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert (request|response)

patch (request|response)

get (response)

update (request|response)

PretargetingConfigDeleteMethodBuilder
Deletes an existing pretargeting config.

PretargetingConfigDimensions
Requests which allow one of these (width, height) pairs will match. All pairs must be supported ad dimensions.

This type is not used in any activity, and only used as part of another schema.

PretargetingConfigExcludedPlacements
Requests containing any of these placements will not match.

This type is not used in any activity, and only used as part of another schema.

PretargetingConfigGetMethodBuilder
Gets a specific pretargeting configuration

PretargetingConfigInsertMethodBuilder
Inserts a new pretargeting configuration.

PretargetingConfigList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

PretargetingConfigListMethodBuilder
Retrieves a list of the authenticated user's pretargeting configurations.

PretargetingConfigMethodsBuilder
A builder providing access to all methods supported on pretargetingConfig resources. It is not used directly, but through the AdExchangeBuyer hub.

PretargetingConfigPatchMethodBuilder
Updates an existing pretargeting config. This method supports patch semantics.

PretargetingConfigPlacements
Requests containing any of these placements will match.

This type is not used in any activity, and only used as part of another schema.

PretargetingConfigUpdateMethodBuilder
Updates an existing pretargeting config.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.