Struct adexchangebuyer1d3::Creative

pub struct Creative {
    pub product_categories: Vec<i32>,
    pub status: Option<String>,
    pub agency_id: Option<String>,
    pub attribute: Vec<i32>,
    pub height: Option<i32>,
    pub advertiser_name: Option<String>,
    pub html_snippet: Option<String>,
    pub buyer_creative_id: Option<String>,
    pub account_id: Option<i32>,
    pub advertiser_id: Vec<String>,
    pub kind: Option<String>,
    pub video_url: Option<String>,
    pub click_through_url: Vec<String>,
    pub corrections: Vec<CreativeCorrections>,
    pub width: Option<i32>,
    pub restricted_categories: Vec<i32>,
    pub vendor_type: Vec<i32>,
    pub filtering_reasons: Option<CreativeFilteringReasons>,
    pub disapproval_reasons: Vec<CreativeDisapprovalReasons>,
    pub sensitive_categories: Vec<i32>,
}

A creative and its classification data.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert creatives (request|response)

list creatives (none)

get creatives (response)

Fields

product_categories

Detected product categories, if any. Read-only. This field should not be set in requests.

status

Creative serving status. Read-only. This field should not be set in requests.

agency_id

The agency id for this creative.

attribute

All attributes for the ads that may be shown from this snippet.

height

Ad height.

advertiser_name

The name of the company being advertised in the creative.

html_snippet

The HTML snippet that displays the ad when inserted in the web page. If set, videoURL should not be set.

buyer_creative_id

A buyer-specific id identifying the creative in this ad.

account_id

Account id.

advertiser_id

Detected advertiser id, if any. Read-only. This field should not be set in requests.

kind

Resource type.

video_url

The url to fetch a video ad. If set, HTMLSnippet should not be set.

click_through_url

The set of destination urls for the snippet.

corrections

Shows any corrections that were applied to this creative. Read-only. This field should not be set in requests.

width

Ad width.

restricted_categories

All restricted categories for the ads that may be shown from this snippet.

vendor_type

All vendor types for the ads that may be shown from this snippet.

filtering_reasons

The filtering reasons for the creative. If this feature is not enabled, please ask your technical account manager. Read-only. This field should not be set in requests.

disapproval_reasons

The reasons for disapproval, if any. Note that not all disapproval reasons may be categorized, so it is possible for the creative to have a status of DISAPPROVED with an empty list for disapproval_reasons. In this case, please reach out to your TAM to help debug the issue. Read-only. This field should not be set in requests.

sensitive_categories

Detected sensitive categories, if any. Read-only. This field should not be set in requests.